requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
pandas>=2.0.0
numpy>=1.24.0
pillow>=10.0.0
matplotlib>=3.7.0
torch>=2.0.0
torchvision>=0.15.0
pyyaml>=6.0